third_party/blink/web_tests/fast/peerconnection/RTCPeerConnection-datachannel-expected.txt

Tests RTCDataChannel.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS dc.send(new Uint8Array(new (new WebAssembly.Memory({ shared:true, initial:0, maximum:0 }).buffer.constructor)())); threw exception TypeError: Failed to execute 'send' on 'RTCDataChannel': The provided ArrayBufferView value must not be shared..
PASS When the data channel send queue is full, send() throws an exception
PASS dc.readyState is "open"
PASS successfullyParsed is true

TEST COMPLETE
